treat cancel action in file dialogs

+++ cnapy/gui_elements/mainwindow.py
@@ -268,8 +268,16 @@ class MainWindow(QMainWindow):
 
     def unsaved_changes(self):
         self.appdata.unsaved = True
+        self.save_project_action.setEnabled(True)
         palette = self.palette()
         palette.setColor(QPalette.Window, Qt.yellow)
+        self.setPalette(palette)
+
+    def nounsaved_changes(self):
+        self.appdata.unsaved = False
+        self.save_project_action.setEnabled(False)
+        palette = self.palette()
+        palette.setColor(QPalette.Window, self.original_color)
         self.setPalette(palette)
 
     def disable_enable_dependent_actions(self):
@@ -329,26 +337,32 @@ class MainWindow(QMainWindow):
     def import_sbml(self):
         dialog = QFileDialog(self)
         filename: str = dialog.getOpenFileName(
-            directory=os.getcwd(), filter="*.xml")
+            directory=os.getcwd(), filter="*.xml")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
 
         self.appdata.project.cobra_py_model = cobra.io.read_sbml_model(
-            filename[0])
+            filename)
         self.centralWidget().update()
 
     @Slot()
     def export_sbml(self):
         dialog = QFileDialog(self)
         filename: str = dialog.getSaveFileName(
-            directory=os.getcwd(), filter="*.xml")
+            directory=os.getcwd(), filter="*.xml")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
 
         cobra.io.write_sbml_model(
-            self.appdata.project.cobra_py_model, filename[0])
+            self.appdata.project.cobra_py_model, filename)
 
     @Slot()
     def load_box_positions(self):
         dialog = QFileDialog(self)
         filename: str = dialog.getOpenFileName(
-            directory=os.getcwd(), filter="*.maps")
+            directory=os.getcwd(), filter="*.maps")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
 
         idx = self.centralWidget().map_tabs.currentIndex()
         if idx < 0:
@@ -356,7 +370,7 @@ class MainWindow(QMainWindow):
             idx = self.centralWidget().map_tabs.currentIndex()
         name = self.centralWidget().map_tabs.tabText(idx)
 
-        with open(filename[0], 'r') as fp:
+        with open(filename, 'r') as fp:
             print(fp)
             self.appdata.project.maps[name]["boxes"] = json.load(fp)
 
@@ -376,9 +390,11 @@ class MainWindow(QMainWindow):
     def load_scenario(self):
         dialog = QFileDialog(self)
         filename: str = dialog.getOpenFileName(
-            directory=os.getcwd(), filter="*.scen")
+            directory=os.getcwd(), filter="*.scen")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
 
-        with open(filename[0], 'r') as fp:
+        with open(filename, 'r') as fp:
             values = json.load(fp)
             self.appdata.project.scen_values.clear()
             for i in values:
@@ -392,9 +408,11 @@ class MainWindow(QMainWindow):
     def load_modes(self):
         dialog = QFileDialog(self)
         filename: str = dialog.getOpenFileName(
-            directory=os.getcwd(), filter="*.modes")
+            directory=os.getcwd(), filter="*.modes")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
 
-        with open(filename[0], 'r') as fp:
+        with open(filename, 'r') as fp:
             self.appdata.project.modes = json.load(fp)
             self.centralWidget().mode_navigator.current = 0
             values = self.appdata.project.modes[0].copy()
@@ -408,12 +426,13 @@ class MainWindow(QMainWindow):
     def change_background(self):
         dialog = QFileDialog(self)
         filename: str = dialog.getOpenFileName(
-            directory=os.getcwd(), filter="*.svg")
+            directory=os.getcwd(), filter="*.svg")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
 
         idx = self.centralWidget().map_tabs.currentIndex()
         name = self.centralWidget().map_tabs.tabText(idx)
-        if filename[0] != '':
-            # try:
+        if filename != '':
             self.appdata.project.maps[name]["background"] = filename[0]
             print(self.appdata.project.maps[name]["background"])
 
@@ -421,8 +440,6 @@ class MainWindow(QMainWindow):
                 self.appdata.project.maps[name]["background"])
             background.setFlags(QGraphicsItem.ItemClipsToShape)
             self.centralWidget().map_tabs.widget(idx).scene.addItem(background)
-            # except:
-            # print("could not update background")
 
             self.centralWidget().update()
             self.centralWidget().map_tabs.setCurrentIndex(idx)
@@ -456,18 +473,22 @@ class MainWindow(QMainWindow):
 
         dialog = QFileDialog(self)
         filename: str = dialog.getSaveFileName(
-            directory=os.getcwd(), filter="*.maps")
+            directory=os.getcwd(), filter="*.maps")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
 
-        with open(filename[0], 'w') as fp:
+        with open(filename, 'w') as fp:
             json.dump(self.appdata.project.maps[name]["boxes"], fp)
 
     @Slot()
     def save_scenario(self):
         dialog = QFileDialog(self)
         filename: str = dialog.getSaveFileName(
-            directory=os.getcwd(), filter="*.scen")
+            directory=os.getcwd(), filter="*.scen")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
 
-        with open(filename[0], 'w') as fp:
+        with open(filename, 'w') as fp:
             json.dump(self.appdata.project.scen_values, fp)
         self.appdata.project.scenario_backup = self.appdata.project.scen_values.copy()
 
@@ -475,9 +496,10 @@ class MainWindow(QMainWindow):
     def save_modes(self):
         dialog = QFileDialog(self)
         filename: str = dialog.getSaveFileName(
-            directory=os.getcwd(), filter="*.modes")
-
-        with open(filename[0], 'w') as fp:
+            directory=os.getcwd(), filter="*.modes")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
+        with open(filename, 'w') as fp:
             json.dump(self.appdata.project.modes, fp)
 
     def reset_scenario(self):
@@ -513,31 +535,37 @@ class MainWindow(QMainWindow):
         self.clear_scenario()
 
         self.setCurrentFile("Untitled project")
-        self.save_project_action.setEnabled(False)
+        self.nounsaved_changes()
 
     @Slot()
     def open_project(self):
         dialog = QFileDialog(self)
         filename: str = dialog.getOpenFileName(
-            directory=os.getcwd(), filter="*.cna")
+            directory=os.getcwd(), filter="*.cna")[0]
+        if not filename or len(filename) == 0 or not os.path.exists(filename):
+            return
 
-        self.appdata.temp_dir = TemporaryDirectory()
+        temp_dir = TemporaryDirectory()
 
-        with ZipFile(filename[0], 'r') as zip_ref:
-            zip_ref.extractall(self.appdata.temp_dir.name)
+        with ZipFile(filename, 'r') as zip_ref:
+            zip_ref.extractall(temp_dir.name)
 
-            with open(self.appdata.temp_dir.name+"/maps.json", 'r') as fp:
-                self.appdata.project.maps = json.load(fp)
+            with open(temp_dir.name+"/maps.json", 'r') as fp:
+                maps = json.load(fp)
 
                 count = 1
-                for name, m in self.appdata.project.maps.items():
-                    m["background"] = self.appdata.temp_dir.name + \
+                for name, m in maps.items():
+                    m["background"] = temp_dir.name + \
                         "/.bg" + str(count) + ".svg"
                     count += 1
 
-            self.appdata.project.cobra_py_model = cobra.io.read_sbml_model(
-                self.appdata.temp_dir.name + "/model.sbml")
+            cobra_py_model = cobra.io.read_sbml_model(
+                temp_dir.name + "/model.sbml")
 
+            self.appdata.temp_dir = temp_dir
+            self.appdata.project.maps = maps
+            self.appdata.project.cobra_py_model = cobra_py_model
+            self.setCurrentFile(filename)
             self.recreate_maps()
             self.centralWidget().mode_navigator.clear()
             self.clear_scenario()
@@ -545,10 +573,8 @@ class MainWindow(QMainWindow):
                 if 'cnapy-default' in r.annotation.keys():
                     self.centralWidget().update_reaction_value(
                         r.id, r.annotation['cnapy-default'])
-            self.save_project_action.setEnabled(True)
+            self.nounsaved_changes()
             self.centralWidget().reaction_list.update()
-
-        self.setCurrentFile(filename[0])
 
     @Slot()
     def save_project(self):
@@ -587,9 +613,7 @@ class MainWindow(QMainWindow):
                     "/.bg" + str(count) + ".svg"
                 count += 1
 
-        palette = self.palette()
-        palette.setColor(QPalette.Window, self.original_color)
-        self.setPalette(palette)
+        self.nounsaved_changes()
 
     @Slot()
     def save_project_as(self):
@@ -600,10 +624,9 @@ class MainWindow(QMainWindow):
 
         if len(filename[0]) != 0:
             self.setCurrentFile(filename[0])
-            self.save_project_action.setEnabled(True)
             self.save_project()
         else:
-            return False
+            return
 
     def recreate_maps(self):
         self.centralWidget().map_tabs.currentChanged.disconnect(self.on_tab_change)
